Sta motoka

+++ database/migrations/2025_09_09_205847_create_lgas_table.php
@@ -11,12 +11,14 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('lgas', function (Blueprint $table) {
-            $table->id();
-            $table->string('name');
-            $table->foreignId('state_id')->constrained('states')->onDelete('cascade');
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('lgas')) {
+            Schema::create('lgas', function (Blueprint $table) {
+                $table->id();
+                $table->string('name');
+                $table->foreignId('state_id')->constrained('states')->onDelete('cascade');
+                $table->timestamps();
+            });
+        }
     }
 
     /**
